VC++ 2008 프로젝트에 x64 플랫폼을 추가함. x64 플랫폼에서 빌드 경고 메시지가 출력되지 않도록 수정함.

+++ KSipServer/CallMap.cpp
@@ -190,7 +190,7 @@ int CCallMap::GetCount()
 	int iCount;
 
 	m_clsMutex.acquire();
-	iCount = m_clsMap.size();
+	iCount = (int)m_clsMap.size();
 	m_clsMutex.release();
 
 	return iCount;
